third_party/WebKit/LayoutTests/editing/input/text-input-controller-leak-document.html

<html>
<body>
    <input id='text' type='text'></input>

    <script src="../../resources/js-test.js"></script>
    <script>
        var input = document.getElementById('text');
        input.focus();
        textInputController.setMarkedText("hello", 0, 5);
        var markedRangeResultGC = internals.observeGC(textInputController.markedRange());
        var selectedRangeResultGC = internals.observeGC(textInputController.selectedRange());
        var firstRectForCharacterRangeGC = internals.observeGC(textInputController.firstRectForCharacterRange(0, 0));
        gc();
        shouldBeTrue('markedRangeResultGC.wasCollected');
        shouldBeTrue('selectedRangeResultGC.wasCollected');
        shouldBeTrue('firstRectForCharacterRangeGC.wasCollected');
    </script>
</body>
</html>
